pyproject.toml
[project]
name = "bitrix-autouser"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniil",email = "fluxxpuff@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.15.0"

